+++ modules/Food_Service/Users/ServeMenus.php
@@ -141,10 +141,9 @@ if ( UserStaffID()
 
 		ListOutput( $RET, $columns, $singular, $plural, [], false, [ 'save' => false, 'search' => false ] );
 
-		$items_RET = DBGet( "SELECT fsi.SHORT_NAME,fsi.DESCRIPTION,fsi.PRICE_STAFF,fsi.ICON
+		$items_RET = DBGet( "SELECT fsi.SHORT_NAME,fsi.DESCRIPTION,fsi.PRICE_STAFF,fsi.ICON,fsmi.MENU_ID
 		FROM food_service_items fsi,food_service_menu_items fsmi
-		WHERE fsmi.MENU_ID='" . (int) $_REQUEST['menu_id'] . "'
-		AND fsi.ITEM_ID=fsmi.ITEM_ID
+		WHERE fsi.ITEM_ID=fsmi.ITEM_ID
 		AND fsmi.CATEGORY_ID IS NOT NULL
 		AND fsi.SCHOOL_ID='" . UserSchool() . "'
 		ORDER BY fsi.SORT_ORDER IS NULL,fsi.SORT_ORDER", [ 'ICON' => 'makeIcon' ], [ 'SHORT_NAME' ] );
@@ -152,7 +151,10 @@ if ( UserStaffID()
 
 		foreach ( (array) $items_RET as $sn => $item )
 		{
-			$items += [ $sn => $item[1]['DESCRIPTION'] ];
+			if ( $item[1]['MENU_ID'] == $_REQUEST['menu_id'] )
+			{
+				$items += [ $sn => $item[1]['DESCRIPTION'] ];
+			}
 		}
 
 		$LO_ret = [ [] ];
@@ -162,7 +164,13 @@ if ( UserStaffID()
 			foreach ( (array) $_SESSION['FSA_sale']['staff_' . UserStaffID() ] as $id => $item_sn )
 			{
 				$price = $items_RET[$item_sn][1]['PRICE_STAFF'];
-				$LO_ret[] = [ 'SALE_ID' => $id, 'PRICE' => $price, 'DESCRIPTION' => $items_RET[$item_sn][1]['DESCRIPTION'], 'ICON' => $items_RET[$item_sn][1]['ICON'] ];
+
+				$LO_ret[] = [
+					'SALE_ID' => $id,
+					'PRICE' => $price,
+					'DESCRIPTION' => $items_RET[$item_sn][1]['DESCRIPTION'],
+					'ICON' => $items_RET[$item_sn][1]['ICON'],
+				];
 			}
 		}