+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 12.8
 ---------------
 - Fix PHP warning undefined array key when Meal Item Short Name was edited in MenuReports.php
 - Fix PHP8.5 deprecated using null as an array offset, use an empty string instead in MenuReports.php & ActivityReport.php
+- Fix PHP warning undefined array key when switching Meals in ServeMenus.php
 
 Changes in 12.7.2
 -----------------